Refactor get-space and create-space into api/spaces
