--- pages/api/spaces/create-new-space.js
@@ -1,33 +1,0 @@
-import dbConnect from '../../../utils/dbConnect';
-import Space from '../../../models/Spaces';
-import User from '../../../models/User';
-
-export default async (req, res) => {
-  const { method } = req;
-
-  if (method !== 'POST') {
-    return;
-  }
-
-  const { name, description, isActive, participants, spaceId } = req.body;
-
-  if (!name || !description) {
-    return res.status(422).json({ message: 'Invalid input' });
-  }
-  await dbConnect();
-
-  try {
-    const space = new Space({
-      name,
-      description,
-      spaceId,
-      participants,
-      isActive,
-    });
-
-    const result = space.save();
-    res.status(201).json({ message: 'Space created!', _id: result._id });
-  } catch (err) {
-    return res.status(500).json({ message: err.message });
-  }
-};
